Moved settings to its own view.
